refactor: split proxy compatibility helpers
