+++ proxy/csswitch_proxy.py
@@ -38,7 +38,10 @@ import urllib.error
 from http.server import BaseHTTPRequestHandler, ThreadingHTTPServer
 
 import dsml_shim
+import model_discovery
+import openai_chat_compat
 import provider_policy
+import responses_compat
 import anthropic_compat
 
 # DSML 兜底 shim 的运行模式：off（默认，字节级透传）/ detect（透传 + 遥测）/ rewrite（改写）。
@@ -364,20 +367,7 @@ def fetch_relay_models():
     if PROV_NAME == "relay":
         headers["anthropic-version"] = "2023-06-01"
     raw = http_get_json(murl, headers)
-    data = raw.get("data") if isinstance(raw, dict) else raw
-    out, ids = [], []
-    for m in data or []:
-        mid = m.get("id") if isinstance(m, dict) else None
-        if not mid:
-            continue
-        ids.append(mid)
-        # 能力位：从上游 supported_parameters 推断，绝不臆测（无该字段 → None）。
-        sp = m.get("supported_parameters") if isinstance(m, dict) else None
-        supports_tools = ("tools" in sp) if isinstance(sp, list) else None
-        out.append({"type": "model", "id": mid,
-                    "display_name": (m.get("display_name") if isinstance(m, dict) else None) or mid,
-                    "supports_tools": supports_tools,
-                    "created_at": "2026-01-01T00:00:00Z"})
+    out, ids = model_discovery.normalize_models_response(raw)
     if ids and PROV_NAME == "relay":
         RELAY_MODELS = ids
     return out
@@ -395,18 +385,12 @@ def build_models_response():
             # force（Science 常驻代理）：只返回一个壳，Science 主列表显示真实模型名。
             # 出站由 resolve_model 的 force 分支覆盖，无需 model_map。app 的 fetch_models
             # 不设 RELAY_FORCE_MODEL，故仍走下面回源拿真实 id 供用户选（两个消费者切分）。
-            shell = [{"type": "model", "id": "claude-opus-4-8",
-                      "display_name": RELAY_FORCE_MODEL, "supports_tools": None,
-                      "created_at": "2026-01-01T00:00:00Z"}]
             log(f"GET /v1/models -> {PROV_NAME}(force 借壳): {RELAY_FORCE_MODEL}")
-            return 200, {"data": shell, "has_more": False,
-                         "first_id": "claude-opus-4-8", "last_id": "claude-opus-4-8"}
+            return model_discovery.force_shell_response(RELAY_FORCE_MODEL)
         try:
             data = fetch_relay_models()
             log(f"GET /v1/models -> {PROV_NAME}(回源): {len(data)} 个模型")
-            return 200, {"data": data, "has_more": False,
-                         "first_id": data[0]["id"] if data else None,
-                         "last_id": data[-1]["id"] if data else None}
+            return model_discovery.live_models_response(data)
         except urllib.error.HTTPError as e:
             detail = ""
             try:
@@ -420,309 +404,56 @@ def build_models_response():
             log(f"GET /v1/models -> {PROV_NAME} 回源网络异常，本地回 502: {e}")
             return 502, {"error_kind": "network", "upstream_status": None, "message": str(e)}
     # 非 relay：静态选择器列表（deepseek/qwen）。
-    data = [{"type": "model", "id": mid, "display_name": disp, "supports_tools": None,
-             "created_at": "2026-01-01T00:00:00Z"} for mid, disp in PROV["models"]]
-    return 200, {"data": data, "has_more": False,
-                 "first_id": data[0]["id"] if data else None,
-                 "last_id": data[-1]["id"] if data else None}
+    return model_discovery.static_models_response(PROV["models"])
 
 
 # ---------- Anthropic -> OpenAI 翻译（qwen 路径） ----------
 def anthropic_to_openai(req):
-    msgs = []
-    sys_prompt = req.get("system")
-    if isinstance(sys_prompt, list):
-        sys_prompt = "\n".join(b.get("text", "") for b in sys_prompt if isinstance(b, dict))
-    if sys_prompt:
-        msgs.append({"role": "system", "content": sys_prompt})
-    for m in req.get("messages", []):
-        role = m.get("role")
-        content = m.get("content")
-        if isinstance(content, str):
-            msgs.append({"role": role, "content": content})
-            continue
-        text_parts, tool_calls, tool_results = [], [], []
-        for blk in content or []:
-            t = blk.get("type")
-            if t == "text":
-                text_parts.append(blk.get("text", ""))
-            elif t == "tool_use":
-                tool_calls.append({
-                    "id": blk.get("id"), "type": "function",
-                    "function": {"name": blk.get("name"),
-                                 "arguments": json.dumps(blk.get("input", {}), ensure_ascii=False)},
-                })
-            elif t == "tool_result":
-                c = blk.get("content")
-                if isinstance(c, list):
-                    c = "".join(x.get("text", "") for x in c if isinstance(x, dict))
-                tool_results.append({"role": "tool", "tool_call_id": blk.get("tool_use_id"),
-                                     "content": c if isinstance(c, str) else json.dumps(c, ensure_ascii=False)})
-        if role == "assistant" and tool_calls:
-            msgs.append({"role": "assistant", "content": "".join(text_parts) or None, "tool_calls": tool_calls})
-        elif tool_results:
-            msgs.extend(tool_results)
-            if text_parts:
-                msgs.append({"role": role, "content": "".join(text_parts)})
-        else:
-            msgs.append({"role": role, "content": "".join(text_parts)})
-    _st = _provider_state(req)
-    out = {"model": provider_policy.resolve_model(req.get("model"), _st),
-           "messages": msgs, "stream": False}
-    if req.get("max_tokens"):
-        out["max_tokens"] = provider_policy.clamp_max_tokens(req["max_tokens"], out["model"], _st)
-    if req.get("temperature") is not None:
-        out["temperature"] = req["temperature"]
-    if req.get("tools"):
-        out["tools"] = [{"type": "function",
-                         "function": {"name": t["name"], "description": t.get("description", ""),
-                                      "parameters": t.get("input_schema", {})}}
-                        for t in req["tools"] if t.get("name")]
-    tcm = map_tool_choice(req.get("tool_choice"), req.get("tools"))
-    if tcm is not None:
-        out["tool_choice"] = tcm
-    if req.get("stop_sequences"):
-        out["stop"] = req["stop_sequences"]
-    if req.get("top_p") is not None:
-        out["top_p"] = req["top_p"]
-    return out
+    return openai_chat_compat.anthropic_to_openai(req, _provider_state(req))
 
 
 def map_tool_choice(tc, tools):
-    """把 Anthropic tool_choice 译成 OpenAI 兼容取值。
-    any 不做通用映射：单工具直接指定该函数（等效强制且不依赖 required）；
-    多工具退 "required"（DashScope 若不支持会以上游错误显式暴露，不静默退化）。"""
-    if not isinstance(tc, dict):
-        return None
-    t = tc.get("type")
-    if t == "auto":
-        return "auto"
-    if t == "none":
-        return "none"
-    if t == "tool" and tc.get("name"):
-        return {"type": "function", "function": {"name": tc["name"]}}
-    if t == "any":
-        names = [x["name"] for x in (tools or []) if x.get("name")]
-        if len(names) == 1:
-            return {"type": "function", "function": {"name": names[0]}}
-        return "required"
-    return None
+    return openai_chat_compat.map_tool_choice(tc, tools)
 
 
 def map_responses_tool_choice(tc, tools):
-    """把 Anthropic tool_choice 译成 Responses API 取值。
-    千问 Responses 在 thinking mode 下会拒绝 required/object 形态；Science 会在
-    reviewing/agent 路径发送强制工具选择。这里保守降级为 auto，宁可不强制工具，
-    也不要让整个请求 400。"""
-    has_tools = bool(tools)
-    if isinstance(tc, str):
-        t = tc
-    elif isinstance(tc, dict):
-        t = tc.get("type")
-    else:
-        t = None
-    if t == "auto":
-        return "auto"
-    if t == "none":
-        return "none"
-    if has_tools:
-        return "auto"
-    return None
+    return responses_compat.map_tool_choice(tc, tools)
 
 
 def responses_max_output_tokens(req, model, state, has_tools):
-    value = provider_policy.clamp_max_tokens(req.get("max_tokens"), model, state)
-    if not value:
-        return value
-    # DashScope Responses accepts 65536 for simple prompts, but real Science tool
-    # requests with very large inbound max_tokens can still fail its compatibility
-    # layer. Keep tool calls on the conservative budget used by the existing Qwen
-    # chat path.
-    if has_tools and "dashscope.aliyuncs.com" in (PROV.get("url") or ""):
-        return min(int(value), 8192)
-    return value
+    return responses_compat.max_output_tokens(req, model, state, has_tools, _is_dashscope_responses())
 
 
 def _is_dashscope_responses():
-    return PROV.get("api_format") == "openai_responses" and "dashscope.aliyuncs.com" in (PROV.get("url") or "")
+    return responses_compat.is_dashscope_responses(PROV)
 
 
 def normalize_responses_tool_parameters(schema):
-    if not isinstance(schema, dict):
-        return {"type": "object", "properties": {}}
-    out = dict(schema)
-    if "properties" in out and not out.get("type"):
-        out["type"] = "object"
-    if out.get("type") != "object":
-        return {"type": "object", "properties": {}}
-    if not isinstance(out.get("properties"), dict):
-        out["properties"] = {}
-    return out
+    return responses_compat.normalize_tool_parameters(schema)
 
 
 def map_responses_tools(tools):
-    out = []
-    for t in tools or []:
-        name = t.get("name")
-        if not name:
-            continue
-        # DashScope Responses currently rejects this function schema in
-        # compatibility mode; omit it rather than failing the entire request.
-        if _is_dashscope_responses() and name == "web_search":
-            continue
-        out.append({
-            "type": "function",
-            "name": name,
-            "description": t.get("description", ""),
-            "parameters": normalize_responses_tool_parameters(t.get("input_schema", {})),
-        })
-    return out
-
-
-def _as_text(value):
-    if isinstance(value, str):
-        return value
-    if isinstance(value, list):
-        return "".join(x.get("text", "") for x in value if isinstance(x, dict))
-    if value is None:
-        return ""
-    return json.dumps(value, ensure_ascii=False)
+    return responses_compat.map_tools(tools, _is_dashscope_responses())
 
 
 def anthropic_to_openai_responses(req):
-    sys_prompt = req.get("system")
-    if isinstance(sys_prompt, list):
-        sys_prompt = "\n".join(b.get("text", "") for b in sys_prompt if isinstance(b, dict))
-
-    items = []
-    for m in req.get("messages", []):
-        role = m.get("role")
-        content = m.get("content")
-        if isinstance(content, str):
-            items.append({"role": role, "content": content})
-            continue
-
-        text_parts = []
-        for blk in content or []:
-            t = blk.get("type")
-            if t == "text":
-                text_parts.append(blk.get("text", ""))
-            elif t == "tool_use":
-                if text_parts:
-                    items.append({"role": role, "content": "".join(text_parts)})
-                    text_parts = []
-                items.append({
-                    "type": "function_call",
-                    "call_id": blk.get("id"),
-                    "name": blk.get("name"),
-                    "arguments": json.dumps(blk.get("input", {}), ensure_ascii=False),
-                })
-            elif t == "tool_result":
-                if text_parts:
-                    items.append({"role": role, "content": "".join(text_parts)})
-                    text_parts = []
-                items.append({
-                    "type": "function_call_output",
-                    "call_id": blk.get("tool_use_id"),
-                    "output": _as_text(blk.get("content")),
-                })
-        if text_parts:
-            items.append({"role": role, "content": "".join(text_parts)})
-
-    _st = _provider_state(req)
-    out = {
-        "model": provider_policy.resolve_model(req.get("model"), _st),
-        "input": items,
-        "stream": False,
-    }
-    if sys_prompt:
-        out["instructions"] = sys_prompt
-    tools = map_responses_tools(req.get("tools"))
-    max_output_tokens = responses_max_output_tokens(req, out["model"], _st, bool(tools))
-    if max_output_tokens:
-        out["max_output_tokens"] = max_output_tokens
-    if req.get("temperature") is not None:
-        out["temperature"] = req["temperature"]
-    if req.get("top_p") is not None:
-        out["top_p"] = req["top_p"]
-    if tools:
-        out["tools"] = tools
-    tcm = map_responses_tool_choice(req.get("tool_choice"), tools)
-    if tcm is not None:
-        out["tool_choice"] = tcm
-    return out
+    return responses_compat.anthropic_to_openai(
+        req,
+        _provider_state(req),
+        _is_dashscope_responses(),
+    )
 
 
 def openai_to_anthropic(resp, model_id):
-    choice = (resp.get("choices") or [{}])[0]
-    msg = choice.get("message", {})
-    blocks = []
-    if msg.get("content"):
-        blocks.append({"type": "text", "text": msg["content"]})
-    for tc in msg.get("tool_calls") or []:
-        fn = tc.get("function", {})
-        try:
-            args = json.loads(fn.get("arguments") or "{}")
-        except Exception:
-            args = {}
-        blocks.append({"type": "tool_use", "id": tc.get("id"), "name": fn.get("name"), "input": args})
-    fr = choice.get("finish_reason")
-    stop = {"stop": "end_turn", "length": "max_tokens", "tool_calls": "tool_use"}.get(fr, "end_turn")
-    usage = resp.get("usage", {})
-    return {
-        "id": resp.get("id", "msg_proxy"), "type": "message", "role": "assistant", "model": model_id,
-        "content": blocks or [{"type": "text", "text": ""}],
-        "stop_reason": stop, "stop_sequence": None,
-        "usage": {"input_tokens": usage.get("prompt_tokens", 0),
-                  "output_tokens": usage.get("completion_tokens", 0)},
-    }
+    return openai_chat_compat.openai_to_anthropic(resp, model_id)
 
 
 def _responses_output_text(item):
-    parts = []
-    for c in item.get("content") or []:
-        if not isinstance(c, dict):
-            continue
-        if c.get("type") in ("output_text", "text"):
-            parts.append(c.get("text", ""))
-    return "".join(parts)
+    return responses_compat.output_text(item)
 
 
 def openai_responses_to_anthropic(resp, model_id):
-    blocks = []
-    for item in resp.get("output") or []:
-        if not isinstance(item, dict):
-            continue
-        t = item.get("type")
-        if t == "message":
-            text = _responses_output_text(item)
-            if text:
-                blocks.append({"type": "text", "text": text})
-        elif t == "function_call":
-            raw_args = item.get("arguments") or "{}"
-            try:
-                args = json.loads(raw_args)
-            except Exception:
-                args = {}
-            blocks.append({
-                "type": "tool_use",
-                "id": item.get("call_id") or item.get("id"),
-                "name": item.get("name"),
-                "input": args,
-            })
-    if not blocks and resp.get("output_text"):
-        blocks.append({"type": "text", "text": resp.get("output_text", "")})
-    usage = resp.get("usage", {})
-    stop = "tool_use" if any(b.get("type") == "tool_use" for b in blocks) else "end_turn"
-    if resp.get("status") == "incomplete":
-        stop = "max_tokens"
-    return {
-        "id": resp.get("id", "msg_proxy"), "type": "message", "role": "assistant", "model": model_id,
-        "content": blocks or [{"type": "text", "text": ""}],
-        "stop_reason": stop, "stop_sequence": None,
-        "usage": {"input_tokens": usage.get("input_tokens", 0),
-                  "output_tokens": usage.get("output_tokens", 0)},
-    }
+    return responses_compat.openai_to_anthropic(resp, model_id)
 
 
 class H(BaseHTTPRequestHandler):
